# packages/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: PersAI Step 2 API Contract
  version: 0.1.0
servers:
  - url: /api/v1
paths:
  /assistant:
    get:
      operationId: getAssistant
      summary: Get current authenticated user assistant lifecycle state
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current assistant state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetAssistantResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Assistant not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    post:
      operationId: postAssistantCreate
      summary: Create assistant for current authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Created assistant state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetAssistantResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Assistant already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /assistant/draft:
    patch:
      operationId: patchAssistantDraft
      summary: Update current authenticated user assistant draft fields
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssistantDraftUpdateRequest"
      responses:
        "200":
          description: Updated assistant state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetAssistantResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Assistant not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /assistant/publish:
    post:
      operationId: postAssistantPublish
      summary: Publish current authenticated user assistant draft to immutable version
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Assistant state with latest published version
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetAssistantResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Assistant not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Concurrent publish conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /assistant/rollback:
    post:
      operationId: postAssistantRollback
      summary: Rollback assistant by publishing a snapshot of a previous version
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssistantRollbackRequest"
      responses:
        "200":
          description: Assistant state with new rolled-back latest published version
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetAssistantResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Assistant or target version not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Rollback conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /assistant/reset:
    post:
      operationId: postAssistantReset
      summary: Reset assistant to a new blank state while preserving attachments
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Assistant state with new reset latest published version
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetAssistantResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Assistant not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /assistant/reapply:
    post:
      operationId: postAssistantReapply
      summary: Reapply latest materialized published assistant spec to runtime
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Assistant state after reapply attempt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetAssistantResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Assistant or published version not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /assistant/runtime/preflight:
    get:
      operationId: getAssistantRuntimePreflight
      summary: Check OpenClaw runtime preflight health/readiness via adapter
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Runtime preflight state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetAssistantRuntimePreflightResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /assistant/chat/web:
    post:
      operationId: postAssistantWebChatTurn
      summary: Send one web chat turn through backend transport and adapter boundary
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssistantWebChatTurnRequest"
      responses:
        "200":
          description: Transport result with canonical chat records and runtime response metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetAssistantWebChatTransportResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Assistant not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Assistant transport unavailable due to lifecycle/apply state or active web chats cap
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /assistant/chat/web/stream:
    post:
      operationId: postAssistantWebChatTurnStream
      summary: Send one web chat turn via streaming-first transport
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssistantWebChatTurnRequest"
      responses:
        "200":
          description: Streaming SSE response with started/delta/runtime_done/completed/interrupted/failed events
          content:
            text/event-stream:
              schema:
                type: string
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Assistant not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Assistant transport unavailable due to lifecycle/apply state or active web chats cap
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /assistant/chats/web:
    get:
      operationId: getAssistantWebChats
      summary: List web chat records for current assistant
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Web chat list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetAssistantWebChatListResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Assistant not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /assistant/chats/web/{chatId}:
    patch:
      operationId: patchAssistantWebChat
      summary: Rename web chat record
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: chatId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssistantWebChatRenameRequest"
      responses:
        "200":
          description: Updated web chat list item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetAssistantWebChatListItemResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Assistant or web chat not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      operationId: deleteAssistantWebChat
      summary: Hard delete web chat with explicit confirmation
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: chatId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssistantWebChatDeleteRequest"
      responses:
        "200":
          description: Hard delete completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteAssistantWebChatResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Assistant or web chat not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /assistant/chats/web/{chatId}/archive:
    post:
      operationId: postAssistantWebChatArchive
      summary: Archive web chat record
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: chatId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Archived web chat list item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetAssistantWebChatListItemResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Assistant or web chat not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /assistant/memory/items:
    get:
      operationId: getAssistantMemoryItems
      summary: List active Memory Center items for the current assistant
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Memory registry items (user-facing summaries, not raw runtime memory)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetAssistantMemoryItemsResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Assistant not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /assistant/memory/items/{itemId}/forget:
    post:
      operationId: postAssistantMemoryItemForget
      summary: Forget (soft-remove) one Memory Center item from the user's list
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: itemId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Item forgotten
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostAssistantMemoryItemForgetResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Assistant or item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /assistant/memory/do-not-remember:
    post:
      operationId: postAssistantMemoryDoNotRemember
      summary: Request to stop treating a web chat turn as remembered; updates control-plane markers
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssistantMemoryDoNotRememberRequest"
      responses:
        "200":
          description: Do-not-remember applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostAssistantMemoryDoNotRememberResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Assistant or messages not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AssistantDraftUpdateRequest:
      type: object
      minProperties: 1
      properties:
        displayName:
          type: string
          nullable: true
        instructions:
          type: string
          nullable: true
    AssistantRollbackRequest:
      type: object
      required:
        - targetVersion
      properties:
        targetVersion:
          type: integer
          minimum: 1
...
    AssistantWebChatTurnRequest:
      type: object
      required:
        - surfaceThreadKey
        - message
      properties:
        surfaceThreadKey:
          type: string
        message:
          type: string
        title:
          type: string
          nullable: true
...
    AssistantWebChatRenameRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          nullable: true
...
    GetAssistantResponse:
      type: object
      required:
        - assistant
      properties:
        requestId:
          type: string
          nullable: true
        assistant:
          $ref: "#/components/schemas/AssistantLifecycleState"
    GetAssistantRuntimePreflightResponse:
      type: object
      required:
        - preflight
      properties:
        requestId:
          type: string
          nullable: true
        preflight:
          $ref: "#/components/schemas/AssistantRuntimePreflightState"
    GetAssistantWebChatTransportResponse:
      type: object
      required:
        - transport
      properties:
        requestId:
          type: string
          nullable: true
        transport:
          $ref: "#/components/schemas/AssistantWebChatTurnState"
    GetAssistantWebChatListResponse:
      type: object
      required:
        - chats
      properties:
        requestId:
          type: string
          nullable: true
        chats:
          type: array
          items:
            $ref: "#/components/schemas/AssistantWebChatListItemState"
    GetAssistantWebChatListItemResponse:
      type: object
      required:
        - chat
      properties:
        requestId:
          type: string
          nullable: true
        chat:
          $ref: "#/components/schemas/AssistantWebChatListItemState"
...
    ErrorEnvelope:
      type: object
      required:
        - error
      properties:
        error:
          $ref: "#/components/schemas/ErrorObject"
        requestId:
          type: string
          nullable: true
...